Introduce public example app, configure JSON build variant via param
