+++ .gitignore
@@ -5,3 +5,4 @@ package_config_subset
 .dart_tool/package_config.json
 .dart_tool/version
 build
+.vscode
