+++ example/lib/theme.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class Palette {
+  static const MaterialColor brandColor = MaterialColor(
+    0xff313A49,
+    <int, Color>{
+      50: Color(0xff313A49),//10%
+      100: Color(0xff313A49),//20%
+      200: Color(0xff313A49),//30%
+      300: Color(0xff313A49),//40%
+      400: Color(0xff313A49),//50%
+      500: Color(0xff313A49),//60%
+      600: Color(0xff313A49),//70%
+      700: Color(0xff313A49),//80%
+      800: Color(0xff313A49),//90%
+      900: Color(0xff313A49),//100%
+    },
+  );
+}
